+++ src/app/modals/HowToPlayModal.tsx
